+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - [6036](https://github.com/vegaprotocol/vega/issues/6036) - Fix `protobuf<->swagger` generation
 - [6248](https://github.com/vegaprotocol/vega/issues/6245) - Candles connection is not returning any candle data
 - [6061](https://github.com/vegaprotocol/vega/issues/6061) - Attempt at stabilizing the tests on the broker in the core
+- [6178](https://github.com/vegaprotocol/vega/issues/6178) - Historical balances fails with `scany` error
 
 ## 0.55.0